# spec/api-openapi.json
{
  "components": {
    "schemas": {
...
      "AuthSessionOut": {
        "properties": {
          "role": {
            "title": "Role",
            "type": "string"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "tier",
          "role"
        ],
        "title": "AuthSessionOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Myink 内部 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/session": {
      "get": {
        "operationId": "auth_session_api_v1_auth_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Auth Session",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}